Se termina disposicion de estudiante
